simulation comments. Moved SimulationJob to a file. Moved some tracing to SimulationJob::run and hid some members.
